refactor(geocoder): styling improvements in response to PR
